+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Changelog
 
 ## Unreleased
+- The `spec:` option now just works with a path to the OpenApi file as well so you don't have to call `OpenapiFirst.load` anymore.
 - Router is optional now.
   You no longer have to add `Router` to your middleware stack. You still can add it to customize behaviour by setting options, but you no longer have to add it.
   If you don't add the Router, make sure you pass `spec:` to your request/response validation middleware.
 
 - Support "4xx" and "4XX" response definitions.
   (4XX is defined in the standard, but 2xx is used in the wild as well 🦁.)
-  
+
 - Removed warning about missing operationId, because operationId is not used until the Responder is used.
 - Raise HandlerNotFoundError when handler cannot be found